+++ scm/models/__init__.py
@@ -3,4 +3,8 @@
 from .address import AddressRequestModel, AddressResponseModel
 from .address_group import AddressGroupRequestModel, AddressGroupResponseModel
 from .application import ApplicationRequestModel, ApplicationResponseModel
+from .application_group import (
+    ApplicationGroupRequestModel,
+    ApplicationGroupResponseModel,
+)
 from .service import ServiceRequestModel, ServiceResponseModel